+++ client_helpers_test.go
