+++ go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/stretchr/testify v1.11.1
 	go.unistack.org/micro-codec-json/v4 v4.1.0
 	go.unistack.org/micro/v4 v4.1.19
-	google.golang.org/protobuf v1.36.8
+	google.golang.org/protobuf v1.36.9
 )
 
 require (
